requirements.txt
# Core
streamlit
python-dotenv
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-text-splitters
pydantic
chromadb
pypdf
sentence-transformers==2.6.1
torch==2.2.2
huggingface-hub
numpy